refactor(#45): extract external links from config file
