+++ packages/web-app/src/conf/Config.js
@@ -1,5 +1,3 @@
-import _ from 'underscore.string';
-
 export const LEFT_TO_RIGHT = 'LTR';
 export const RIGHT_TO_LEFT = 'RTL';
 
@@ -43,107 +41,6 @@ export const ADVANCED_SEARCH_TYPES = {
   ENTRANCES: 'entrances',
   MASSIFS: 'massifs',
   ORGANIZATIONS: 'grottos'
-};
-
-function generateLinks(link, defaultLang) {
-  const resultArray = {};
-  Object.keys(AVAILABLE_LANGUAGES).forEach(value => {
-    resultArray[value] = _.replaceAll(link, '%s', _.capitalize(value));
-  });
-  resultArray['*'] = _.replaceAll(link, '%s', _.capitalize(defaultLang));
-  return resultArray;
-}
-
-// ===== Misc links
-export const bbsLink = {
-  '*': 'https://www.ssslib.ch/bbs/'
-};
-
-export const wikicavesLink = {
-  '*': 'http://www.wikicaves.org/'
-};
-
-export const contributorsLink = {
-  '*': 'https://wiki.grottocenter.org/wiki/GrottoCenter:Contributors'
-};
-
-export const facebookLink = {
-  '*': 'https://www.facebook.com/GrottoCenter'
-};
-
-export const twitterLink = {
-  '*': 'https://twitter.com/grottocenter'
-};
-
-export const githubLink = {
-  '*': 'https://github.com/GrottoCenter'
-};
-
-export const licenceLinks = {
-  fr: 'https://creativecommons.org/licenses/by-sa/3.0/fr/',
-  es: 'https://creativecommons.org/licenses/by-sa/3.0/deed.es_ES',
-  ca: 'https://creativecommons.org/licenses/by-sa/3.0/deed.ca',
-  de: 'https://creativecommons.org/licenses/by-sa/3.0/deed.de',
-  pt: 'https://creativecommons.org/licenses/by-sa/3.0/deed.pt_PT',
-  nl: 'https://creativecommons.org/licenses/by-sa/3.0/deed.nl',
-  it: 'https://creativecommons.org/licenses/by-sa/3.0/deed.it',
-  '*': 'https://creativecommons.org/licenses/by-sa/3.0/'
-};
-
-export const licensesODBLink = {
-  '*': 'https://opendatacommons.org/licenses/odbl/'
-};
-
-export const contactLinks = {
-  fr: ' http://fr.wikicaves.org/contact',
-  '*': 'http://en.wikicaves.org/contact'
-};
-
-export const fseLinks = {
-  fr: 'http://eurospeleo.eu/fr/',
-  '*': 'http://eurospeleo.eu/en/'
-};
-
-export const userguideLinks = {
-  fr: 'https://wiki.grottocenter.org/wiki/GrottoCenter:Fr/User_Guide',
-  '*': 'https://wiki.grottocenter.org/wiki/GrottoCenter:En/User_Guide'
-};
-
-export const wikiBatsLinks = generateLinks(
-  'https://wiki.grottocenter.org/wiki/GrottoCenter:%s/bats',
-  'en'
-);
-export const wikiBBSLinks = generateLinks(
-  'https://wiki.grottocenter.org/wiki/GrottoCenter:%s/bbs',
-  'en'
-);
-export const legalLinks = generateLinks(
-  'https://wiki.grottocenter.org/wiki/GrottoCenter:%s/Legal_and_Privacy_Statement',
-  'en'
-);
-
-export const pftGdLink =
-  'https://docs.google.com/document/d/1SccuusPQcxrZJI3nvWcbUc2dgGyKc4ZJXqQzSPeE9Hg/edit?usp=sharing';
-
-export const contributeLinks = {
-  fr: 'http://fr.wikicaves.org/contribute-participer',
-  '*': 'http://en.wikicaves.org/contribute-participer'
-};
-
-export const restApiLinks = {
-  fr: 'https://fr.wikipedia.org/wiki/Representational_state_transfer',
-  '*': 'https://en.wikipedia.org/wiki/Representational_state_transfer'
-};
-
-export const wikiApiLinks = {
-  fr: 'https://fr.wikipedia.org/wiki/Interface_de_programmation',
-  '*': 'https://en.wikipedia.org/wiki/Application_programming_interface'
-};
-
-// ===== Blogger
-export const bloggerLinks = {
-  fr: 'http://blog-fr.grottocenter.org/',
-  '*': 'http://blog-en.grottocenter.org/'
 };
 
 export const bloggerIcons = {
